requirements.txt
openslide-python
Pillow
clarifai_grpc